requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
sqlalchemy
psycopg2-binary
groq
pyjwt
gunicorn